requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
joblib==1.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
sentence-transformers==2.3.1
torch==2.1.2
matplotlib==3.8.2
seaborn==0.13.1

# Data Pipeline Dependencies
pdfminer.six==20231228
pytesseract==0.3.10
Pillow==10.2.0
boto3==1.34.34
botocore==1.34.34
tqdm==4.66.1
pdf2image==1.17.0
python-dateutil==2.8.2
regex==2023.12.25
